static cast

+++ prim_mst.cpp
@@ -134,7 +134,7 @@ void MST::Prim(Graph graph) {
     // for each v search edge list
     // find smallest edge for that v
     pqueue.Push(dist[i], i);
-    unsigned int prev = inf;
+    unsigned int prev = static_cast<unsigned int>(inf);
 
     while (pqueue.Size() > 0) {
       // get destination(vertex) w/ smallest weight
